Move item search configs into app/search/items
