+++ src/app/dim-api/reducer.ts
@@ -16,8 +16,8 @@ import { t } from 'app/i18next-t';
 import { convertDimLoadoutToApiLoadout } from 'app/loadout/loadout-type-converters';
 import { recentSearchComparator } from 'app/search/autocomplete';
 import { CUSTOM_TOTAL_STAT_HASH } from 'app/search/d2-known-values';
-import { FilterContext } from 'app/search/item-filter-types';
-import { buildItemFiltersMap } from 'app/search/item-search-filter';
+import { FilterContext } from 'app/search/items/item-filter-types';
+import { buildItemFiltersMap } from 'app/search/items/item-search-filter';
 import { parseAndValidateQuery } from 'app/search/search-filter';
 import { count, uniqBy } from 'app/utils/collections';
 import { emptyArray } from 'app/utils/empty';
