+++ src/app/inventory/spreadsheets.ts
@@ -2,8 +2,8 @@ import { currentAccountSelector } from 'app/accounts/selectors';
 import { gaEvent } from 'app/google';
 import { LoadoutsByItem, loadoutsByItemSelector } from 'app/loadout/selectors';
 import { D1_StatHashes } from 'app/search/d1-known-values';
+import D2Sources from 'app/search/items/search-filters/d2-sources';
 import { dimArmorStatHashByName } from 'app/search/search-filter-values';
-import D2Sources from 'app/search/search-filters/d2-sources';
 import { ThunkResult } from 'app/store/types';
 import { filterMap } from 'app/utils/collections';
 import { compareBy } from 'app/utils/comparators';
