+++ src/app/item-picker/ItemPicker.tsx
@@ -1,7 +1,7 @@
 import ClassIcon from 'app/dim-ui/ClassIcon';
 import { t } from 'app/i18next-t';
 import SearchBar from 'app/search/SearchBar';
-import { filterFactorySelector } from 'app/search/item-search-filter';
+import { filterFactorySelector } from 'app/search/items/item-search-filter';
 import { uniqBy } from 'app/utils/collections';
 import clsx from 'clsx';
 import { BucketHashes } from 'data/d2/generated-enums';
