+++ src/app/item-triage/ItemTriage.tsx
@@ -16,8 +16,8 @@ import { editLoadout } from 'app/loadout-drawer/loadout-events';
 import InGameLoadoutIcon from 'app/loadout/ingame/InGameLoadoutIcon';
 import { isInGameLoadout } from 'app/loadout/loadout-types';
 import { loadoutsByItemSelector } from 'app/loadout/selectors';
-import { filterFactorySelector } from 'app/search/item-search-filter';
-import { loadoutToSearchString } from 'app/search/search-filters/loadouts';
+import { filterFactorySelector } from 'app/search/items/item-search-filter';
+import { loadoutToSearchString } from 'app/search/items/search-filters/loadouts';
 import { AppIcon, compareIcon, editIcon } from 'app/shell/icons';
 import WishListPerkThumb from 'app/wishlists/WishListPerkThumb';
 import { wishListSelector } from 'app/wishlists/selectors';
