+++ src/app/item-triage/triage-utils.ts
@@ -3,8 +3,8 @@ import { DimItem } from 'app/inventory/item-types';
 import { keyByStatHash, StatLookup } from 'app/inventory/store/stats';
 import { armorStats, CUSTOM_TOTAL_STAT_HASH, TOTAL_STAT_HASH } from 'app/search/d2-known-values';
 import { ItemFilter } from 'app/search/filter-types';
+import { classFilter, itemTypeFilter } from 'app/search/items/search-filters/known-values';
 import { quoteFilterString } from 'app/search/query-parser';
-import { classFilter, itemTypeFilter } from 'app/search/search-filters/known-values';
 import { getInterestingSocketMetadatas, isArtifice, isClassCompatible } from 'app/utils/item-utils';
 import { getIntrinsicArmorPerkSocket } from 'app/utils/socket-utils';
 import _ from 'lodash';
