+++ src/app/organizer/ItemTable.tsx
@@ -26,7 +26,7 @@ import { convertToLoadoutItem, newLoadout } from 'app/loadout-drawer/loadout-uti
 import { loadoutsByItemSelector } from 'app/loadout/selectors';
 import { useD2Definitions } from 'app/manifest/selectors';
 import { showNotification } from 'app/notifications/notifications';
-import { searchFilterSelector } from 'app/search/item-search-filter';
+import { searchFilterSelector } from 'app/search/items/item-search-filter';
 import { setSettingAction } from 'app/settings/actions';
 import { toggleSearchQueryComponent } from 'app/shell/actions';
 import { AppIcon, faCaretDown, faCaretUp, spreadsheetIcon, uploadIcon } from 'app/shell/icons';
