+++ src/app/search/SearchBar.tsx
@@ -51,7 +51,7 @@ import HighlightedText from './HighlightedText';
 import styles from './SearchBar.m.scss';
 import { buildArmoryIndex } from './armory-search';
 import createAutocompleter, { SearchItem, SearchItemType } from './autocomplete';
-import { searchConfigSelector, validateQuerySelector } from './item-search-filter';
+import { searchConfigSelector, validateQuerySelector } from './items/item-search-filter';
 import {
   loadoutSearchConfigSelector,
   validateLoadoutQuerySelector,
